tests/requirements-test.txt
pytest==7.4.3
pytest-mock==3.12.0
pytest-cov==4.1.0
python-dotenv==1.0.0
codaio==0.6.12 